refactor m.playblast module over multiple files instead of one giant file. Clean up and document code
