+++ pipeline/pipe/glui/dialogs.py
@@ -23,7 +23,9 @@ class ButtonPair:
         ok_name: str = "OK",
         cancel_name: str = "Cancel",
     ) -> None:
-        button_list = QtWidgets.QDialogButtonBox.StandardButtons(QtWidgets.QDialogButtonBox.Ok)
+        button_list = QtWidgets.QDialogButtonBox.StandardButtons(
+            QtWidgets.QDialogButtonBox.Ok
+        )
         if has_cancel_button:
             button_list |= QtWidgets.QDialogButtonBox.Cancel
         self.buttons = QtWidgets.QDialogButtonBox(button_list)
